REFACTOR
