+++ lib/chrome/webdriver/chromeDelegate.js
@@ -3,29 +3,19 @@
 const log = require('intel').getLogger('browsertime.chrome');
 const { promisify } = require('util');
 const { Type } = require('selenium-webdriver').logging;
-const get = require('lodash.get');
-const harBuilder = require('../../support/har/');
-const perflogParser = require('chrome-har');
+const cpuMetrics = require('./cpuMetrics');
 const fs = require('fs');
+const path = require('path');
 const { parseCpuTrace } = require('../cpu');
+const speedline = require('./speedline');
+const har = require('./har');
+const harBuilder = require('../../support/har/');
 const webdriver = require('selenium-webdriver');
-const images = require('../../support/images');
 const traceCategoriesParser = require('../traceCategoriesParser');
 const pathToFolder = require('../../support/pathToFolder');
-const path = require('path');
-const speedline = require('speedline-core');
-const visualMetricsExtra = require('../../video/postprocessing/visualmetrics/extraMetrics');
 const ChromeDevtoolsProtocol = require('./chromeDevtoolsProtocol');
 const { Android } = require('../../android');
 const unlink = promisify(fs.unlink);
-
-function pad(n) {
-  const extraZeroes = 6 - n.toString().length;
-  for (let i = 0; i < extraZeroes; i++) {
-    n = '0' + n;
-  }
-  return n;
-}
 
 class ChromeDelegate {
   constructor(storageManager, options) {
@@ -42,7 +32,7 @@ class ChromeDelegate {
     this.isTracing = false;
   }
 
-  async onStart() {
+  async beforeBrowserStart() {
     this.hars = [];
     this.events = [];
     if (this.chrome.android) {
@@ -54,14 +44,14 @@ class ChromeDelegate {
   }
 
   async afterBrowserStart() {
-    // We create the cdp as early as possibble so it can be used in scripts
+    // We create the cdp as early as possible so it can be used in scripts
     // https://bugs.chromium.org/p/chromium/issues/detail?id=824626
     // https://github.com/cyrus-and/chrome-remote-interface/issues/332
     this.cdpClient = new ChromeDevtoolsProtocol(this.options);
     return this.cdpClient.setup();
   }
 
-  async onStartIteration(runner) {
+  async beforeStartIteration(runner) {
     if (this.options.injectJs) {
       await this.cdpClient.injectJavaScript(this.options.injectJs);
     }
@@ -103,27 +93,24 @@ class ChromeDelegate {
     // and since Chromedriver 2.29 there's a bug
     // https://bugs.chromium.org/p/chromedriver/issues/detail?id=1811
     // that can be fixed by emptying the logs twice :|
-    await runner.getLogs(Type.PERFORMANCE);
-    await runner.getLogs(Type.PERFORMANCE);
-
-    if (this.skipHar) {
-      return;
-    }
+    // await runner.getLogs(Type.PERFORMANCE);
+    // await runner.getLogs(Type.PERFORMANCE);
   }
 
+  /*
   async clear(runner) {
     await runner.getLogs(Type.PERFORMANCE);
     await runner.getLogs(Type.PERFORMANCE);
   }
-
-  async onStopIteration() {
-    if (this.chrome.collectNetLog && !this.chrome.android) {
-      const netlog = `${this.baseDir}/chromeNetlog.json`;
-      return unlink(netlog);
-    }
-  }
+  */
 
   async beforeEachURL(runner, url) {
+    // remove irrelevant entries from performance log
+    // and since Chromedriver 2.29 there's a bug
+    // https://bugs.chromium.org/p/chromedriver/issues/detail?id=1811s
+    await runner.getLogs(Type.PERFORMANCE);
+    await runner.getLogs(Type.PERFORMANCE);
+
     if (this.options.cookie && url) {
       await this.cdpClient.setCookies(url, this.options.cookie);
     } else if (this.options.cookie) {
@@ -140,18 +127,25 @@ class ChromeDelegate {
     }
   }
 
-  async beforeCollect() {
-    if (this.collectTracingEvents) {
+  async afterPageCompleteCheck(runner, index) {
+    const result = {};
+    if (this.collectTracingEvents && this.isTracing) {
       // We are ready and can stop collecting events
       this.isTracing = false;
       this.events = await this.cdpClient.stopTrace();
     }
-  }
 
-  async onCollect(runner, index, result) {
     if (this.android && this.options.androidPower) {
       result.power = await this.android.measurePowerUsage(
         this.chrome.android.package
+      );
+    }
+
+    if (this.options.verbose >= 2 || this.chrome.enableChromeDriverLog) {
+      return this.storageManager.gzip(
+        `${this.baseDir}/chromedriver.log`,
+        `${this.baseDir}/chromedriver-${index}.log.gz`,
+        true
       );
     }
 
@@ -176,8 +170,23 @@ class ChromeDelegate {
       );
     }
 
-    const rawCDPMetrics = await this.cdpClient.getPerformanceMetrics();
+    if (!this.skipHar) {
+      result.har = await har(
+        runner,
+        result,
+        index,
+        this.cdpClient,
+        this.logPerfEntries,
+        this.chrome.includeResponseBodies,
+        this.chrome.mobileEmulation,
+        this.android
+      );
+    }
+    return result;
+  }
 
+  async afterEachURL(runner, index, result) {
+    const rawCDPMetrics = await this.cdpClient.getPerformanceMetrics();
     const cleanedMetrics = {};
     let ns, fmp;
     for (let m of rawCDPMetrics.metrics) {
@@ -210,18 +219,6 @@ class ChromeDelegate {
       );
     }
 
-    if (this.skipHar) {
-      return;
-    }
-
-    log.debug('Getting performance logs from Chrome');
-
-    const logs = await runner.getLogs(Type.PERFORMANCE);
-    const messages = logs.map(entry => JSON.parse(entry.message).message);
-
-    if (this.logPerfEntries) {
-      result.extraJson[`chromePerflog-${index}.json`] = messages;
-    }
     if (this.collectTracingEvents) {
       const trace = traceCategoriesParser.parse(this.events, result.url);
       result.extraJson[`trace-${index}.json`] = trace;
@@ -233,68 +230,7 @@ class ChromeDelegate {
         this.chrome.visualMetricsUsingTrace &&
         this.options.cpu
       ) {
-        try {
-          const navStart = trace.traceEvents.filter(
-            task =>
-              task.cat === 'blink.user_timing' &&
-              task.name === 'navigationStart'
-          );
-          navStart.sort(function(a, b) {
-            return a.ts - b.ts;
-          });
-          log.debug('Get Speedline result from the trace');
-          const speedlineResult = await speedline(trace.traceEvents, {
-            timeOrigin: navStart[0].ts,
-            fastMode: true
-          });
-          log.debug('Got Speedline result.');
-          const startTs = speedlineResult.beginning;
-          const visualProgress = speedlineResult.frames
-            .map(frame => {
-              const ts = Math.floor(frame.getTimeStamp() - startTs);
-              return `${ts}=${Math.floor(frame.getProgress())}%`;
-            })
-            .join(', ');
-
-          result.visualMetrics = visualMetricsExtra({
-            FirstVisualChange: Number(speedlineResult.first.toFixed(0)),
-            LastVisualChange: Number(speedlineResult.complete.toFixed(0)),
-            SpeedIndex: Number(speedlineResult.speedIndex.toFixed(0)),
-            PerceptualSpeedIndex: Number(
-              speedlineResult.perceptualSpeedIndex.toFixed(0)
-            ),
-            VisualProgress: visualProgress
-          }).visualMetrics;
-
-          const promises = [];
-          for (let frame of speedlineResult.frames) {
-            // follow the name standard of Visual Metrics
-            // ms_000000.jpg
-            const d = new Date(frame.getTimeStamp() - startTs);
-            const name = 'ms_' + pad(d.getTime());
-            promises.push(
-              images.saveJpg(
-                name,
-                frame.getImage(),
-                result.url,
-                this.storageManager,
-                {
-                  type: 'jpg',
-                  jpg: {
-                    quality: 80
-                  },
-                  maxSize: 400
-                },
-                path.join('filmstrip', index + ''),
-                this.options
-              )
-            );
-          }
-
-          await Promise.all(promises);
-        } catch (e) {
-          log.error('Could not generate Visual Metrics using SpeedLine', e);
-        }
+        result.visualMetrics = speedline(trace, result, index);
       }
     }
 
@@ -303,157 +239,34 @@ class ChromeDelegate {
       result.browserScripts.pageinfo &&
       result.browserScripts.pageinfo.longTask
     ) {
-      let totalDurationFirstPaint = 0;
-      let totalDurationFirstContentFulPaint = 0;
-      let totalDurationAfterLoadEventEnd = 0;
-      let totalDuration = 0;
-
-      // https://web.dev/lighthouse-max-potential-fid/
-      let maxPotentialFid = 0;
-      // https://web.dev/tbt/ - total blocking time
-      let totalBlockingTime = 0;
-
-      const firstPaint = get(
-        result,
-        'browserScripts.timings.paintTiming.first-paint'
-      );
-      const firstContentfulPaint = get(
-        result,
-        'browserScripts.timings.paintTiming.first-contentful-paint'
-      );
-
-      const loadEventEnd = get(result, 'browserScripts.timings.loadEventEnd');
-
-      let longTasksBeforeFirstPaint = 0;
-      let longTasksBeforeFirstContentfulPaint = 0;
-      let longTasksAfterLoadEventEnd = 0;
-
-      for (let longTask of result.browserScripts.pageinfo.longTask) {
-        totalDuration += longTask.duration;
-        if (firstPaint && longTask.startTime < firstPaint) {
-          longTasksBeforeFirstPaint++;
-          totalDurationFirstPaint += longTask.duration;
-        }
-        if (firstContentfulPaint && longTask.startTime < firstContentfulPaint) {
-          longTasksBeforeFirstContentfulPaint++;
-          totalDurationFirstContentFulPaint += longTask.duration;
-        }
-        if (
-          firstContentfulPaint &&
-          longTask.startTime > firstContentfulPaint &&
-          longTask.duration > maxPotentialFid
-        ) {
-          maxPotentialFid = longTask.duration;
-        }
-        if (firstContentfulPaint && longTask.startTime > firstContentfulPaint) {
-          totalBlockingTime +=
-            longTask.duration - (this.options.minLongTaskLength || 50);
-        }
-        if (loadEventEnd && longTask.startTime > loadEventEnd) {
-          longTasksAfterLoadEventEnd++;
-          totalDurationAfterLoadEventEnd += longTask.duration;
-        }
-      }
-
-      const cpu = {
-        longTasks: {
-          tasks: result.browserScripts.pageinfo.longTask.length,
-          totalDuration,
-          totalBlockingTime: Number(totalBlockingTime.toFixed(0)),
-          maxPotentialFid: Number(maxPotentialFid.toFixed(0)),
-          beforeFirstPaint: {
-            tasks: longTasksBeforeFirstPaint,
-            totalDuration: totalDurationFirstPaint
-          },
-          beforeFirstContentfulPaint: {
-            tasks: longTasksBeforeFirstContentfulPaint,
-            totalDuration: totalDurationFirstContentFulPaint
-          },
-          afterLoadEventEnd: {
-            tasks: longTasksAfterLoadEventEnd,
-            totalDuration: totalDurationAfterLoadEventEnd
-          }
-        }
-      };
-
+      const cpuData = cpuMetrics(result);
       if (result.cpu) {
-        result.cpu.longTasks = cpu.longTasks;
+        result.cpu.longTasks = cpuData.longTasks;
       } else {
-        result.cpu = cpu;
+        result.cpu = cpuData;
       }
     }
-
-    // CLEANUP since Chromedriver 2.29 there's a bug
-    // https://bugs.chromium.org/p/chromedriver/issues/detail?id=1811
-    await runner.getLogs(Type.PERFORMANCE);
-
-    const har = perflogParser.harFromMessages(messages);
-
-    if (
-      this.chrome.includeResponseBodies === 'html' ||
-      this.chrome.includeResponseBodies === 'all'
-    ) {
-      await this.cdpClient.setResponseBodies(har);
-    }
-
-    const versionInfo = (await this.cdpClient.send(
-      'Browser.getVersion'
-    )).product.split('/');
-    const info = {
-      name: versionInfo[0],
-      version: versionInfo[1]
-    };
-
-    if (this.chrome.mobileEmulation) {
-      info.name = `Chrome Emulated ${this.chrome.mobileEmulation.deviceName}`;
-    }
-    harBuilder.addBrowser(har, info.name, info.version);
-
-    if (this.android) {
-      har.log._android = await this.android.getModel();
-    }
-
-    if (har.log.pages.length > 0) {
-      har.log.pages[0].title = `${result.url} run ${index}`;
-      // Hack to add the URL from a SPA
-      if (result.alias && !this.aliasAndUrl[result.alias]) {
-        this.aliasAndUrl[result.alias] = result.url;
-        har.log.pages[0]._url = result.url;
-      } else if (result.alias && this.aliasAndUrl[result.alias]) {
-        har.log.pages[0]._url = this.aliasAndUrl[result.alias];
-      } else {
-        har.log.pages[0]._url = result.url;
-      }
-    }
-
-    this.hars.push(har);
   }
 
-  async afterCollect(index) {
-    if (this.options.verbose >= 2 || this.chrome.enableChromeDriverLog) {
-      return this.storageManager.gzip(
-        `${this.baseDir}/chromedriver.log`,
-        `${this.baseDir}/chromedriver-${index}.log.gz`,
-        true
-      );
-    }
+  getCDPClient() {
+    return this.cdpClient;
   }
 
-  async sendAndGetDevToolsCommand(cmd, params = {}) {
-    return this.cdpClient.send(cmd, params);
-  }
-  async sendDevToolsCommand(cmd, params = {}) {
-    return this.cdpClient.send(cmd, params);
-  }
-
+  /*
   failing(url) {
     if (this.skipHar) {
       return;
     }
     this.hars.push(harBuilder.getEmptyHAR(url, this.options.browser));
   }
+  */
 
-  async onStop() {
+  async beforeBrowserStop() {
+    if (this.chrome.collectNetLog && !this.chrome.android) {
+      const netlog = `${this.baseDir}/chromeNetlog.json`;
+      return unlink(netlog);
+    }
+
     if (this.android && this.options.androidPower) {
       await this.android.stopPowerTesting();
     }
@@ -464,6 +277,9 @@ class ChromeDelegate {
     if (this.chrome.android) {
       await this.android.removeFw();
     }
+  }
+
+  async getHARs() {
     if (!this.skipHar) {
       return { har: harBuilder.mergeHars(this.hars) };
     } else return {};
