+++ lib/core/engine/command/chromeDevToolsProtocol.js
@@ -11,10 +11,8 @@ class DevToolsProtocol {
   async sendAndGet(command, args) {
     if (this.browserName === 'chrome' || this.browserName === 'edge') {
       try {
-        const result = await this.engineDelegate.sendAndGetDevToolsCommand(
-          command,
-          args
-        );
+        const client = this.engineDelegate.getCDPClient();
+        const result = await client.send(command, args);
         return result;
       } catch (e) {
         log.error(
@@ -35,7 +33,8 @@ class DevToolsProtocol {
   async send(command, args) {
     if (this.browserName === 'chrome' || this.browserName === 'edge') {
       try {
-        return this.engineDelegate.sendDevToolsCommand(command, args);
+        const client = this.engineDelegate.getCDPClient();
+        return client.send(command, args);
       } catch (e) {
         log.error(
           'Could not send to DevToolsProtocol command %s with args %j',
