+++ lib/core/engine/command/measure.js
@@ -120,7 +120,7 @@ class Measure {
       await this.extensionServer.setup(url, this.browser, this.options);
     }
     if (this.numberOfVisitedPages === 0) {
-      await this.engineDelegate.onStartIteration(this.browser, this.index);
+      await this.engineDelegate.beforeStartIteration(this.browser, this.index);
     }
     this.numberOfVisitedPages++;
     return this.browser.loadAndWait(url, this.pageCompleteCheck);
@@ -155,7 +155,7 @@ class Measure {
       await this.extensionServer.setup(url, this.browser, this.options);
     }
     if (this.numberOfVisitedPages === 0) {
-      await this.engineDelegate.onStartIteration(this.browser, this.index);
+      await this.engineDelegate.beforeStartIteration(this.browser, this.index);
     }
 
     this.result.push(getNewResult());
@@ -188,7 +188,7 @@ class Measure {
       }
     }
 
-    await this.engineDelegate.clear(this.browser);
+    // await this.engineDelegate.clear(this.browser);
 
     if (this.options.tcpdump) {
       await this.tcpDump.start(this.index);
@@ -199,6 +199,14 @@ class Measure {
       try {
         await this.engineDelegate.beforeEachURL(this.browser, url);
         await this.browser.loadAndWait(url, this.pageCompleteCheck);
+        const res = await this.engineDelegate.afterPageCompleteCheck(
+          this.browser,
+          url
+        );
+        this.result[this.numberOfMeasuredPages] = merge(
+          this.result[this.numberOfMeasuredPages],
+          res
+        );
       } catch (e) {
         this.result[this.numberOfMeasuredPages].error = [e.name];
         // The page failed loading but make sure we keep track of
@@ -220,7 +228,15 @@ class Measure {
    * @returns {Promise} Promise object represents all the metrics has been collected.
    */
   async stop() {
-    log.verbose('Stop measuring');
+    log.debug('Stop measuring');
+    const res = await this.engineDelegate.afterPageCompleteCheck(
+      this.browser,
+      url
+    );
+    this.result[this.numberOfMeasuredPages] = merge(
+      this.result[this.numberOfMeasuredPages],
+      res
+    );
     return this.collect();
   }
 
@@ -257,7 +273,7 @@ class Measure {
       await this.tcpDump.stop();
     }
     // This stops collecting trace logs in Chrome etc
-    await this.engineDelegate.beforeCollect();
+    // await this.engineDelegate.beforeCollect();
 
     log.verbose('Collecting metrics');
     // Collect all the metrics through JavaScript
@@ -375,7 +391,7 @@ class Measure {
       this.tcpDump.mv(url, this.index);
     }
 
-    await this.engineDelegate.onCollect(
+    await this.engineDelegate.afterEachURL(
       this.browser,
       this.index,
       this.result[this.numberOfMeasuredPages]
