+++ lib/core/engine/index.js
@@ -152,7 +152,7 @@ class Engine {
     );
 
     await storageManager.createDataDir();
-    await engineDelegate.onStart(name, options);
+    await engineDelegate.beforeBrowserStart(name, options);
     const collector = new Collector(name, storageManager, options);
 
     if (this.android) {
@@ -189,7 +189,7 @@ class Engine {
       }
     }
 
-    const extras = await engineDelegate.onStop();
+    const extras = await engineDelegate.getHARs();
 
     // Backfill the fully loaded data that we extract from the HAR
     if (extras.har) {
