+++ lib/core/engine/iteration.js
@@ -201,7 +201,6 @@ class Iteration {
         );
         await commands.measure.stop();
       }
-      await this.engineDelegate.onStopIteration(browser, index, result);
 
       for (const postScript of this.postScripts) {
         await postScript(context, commands);
@@ -213,6 +212,8 @@ class Iteration {
           result.length > 0 ? result[0].url : 'https://debug.sitespeed.io';
         await this.video.stop(url);
       }
+
+      await this.engineDelegate.beforeBrowserStop(browser, index, result);
       await browser.stop();
 
       if (isAndroidConfigured(options)) {
@@ -259,8 +260,6 @@ class Iteration {
       if (isAndroidConfigured(options)) {
         result.batteryTemperature = batteryTemperature;
       }
-
-      await this.engineDelegate.afterCollect(index, result);
 
       if (commands.meta.description) {
         result.description = commands.meta.description;
